% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReFernment.R
\name{ReFernment}
\alias{ReFernment}
\title{Annotate RNA editing in the GB files of plastomes}
\usage{
ReFernment(gbFolderPath, gffFolderPath, outFolderPath, genomes)
}
\arguments{
\item{gbFolderPath}{this is the path to the folder containing the gb file(s)
that you would like annotated.}

\item{gffFolderPath}{this is the path to the folder containing the gff 
file(s) corresponding to the gff files you would like annotated.}

\item{outFolderPath}{this is the path to the folder where the new genome
annotatons will be produced. The output is in gb format and the new 
annotations will be in the feature table. It is importatnt to note that you
should have the output file seperate from your input files, as this prevents
the script from overwriting the original files!}

\item{genomes}{this should be a list containing the filenames of all the 
genomes that you would like to annotate. Note that you should not include
the file endings of these files.}
}
\value{
an annotated gb file. This file will have the required misc_feature
annotations as well as a translation corrected for RNA editing.
}
\description{
annotates the GB files of plastomes with high levels of RNA
editing. Provides annotations to satisfy GenBank submission requirements. 
Takes as input the paths to the folders containing required file inputs
(a gff3 annotation and .gb file), the desired output path and the names of 
the files that the user wishes to annotate. Each filetype needn't be in 
seperate folders, but it is reccomended. This is especially true for output 
file path, because if it isn't sepereate from input gb files, the original 
files will be overwritten! Be sure that the file names for a respective 
genome are the same
across file types!
}
\examples{
genomes <- c("Asplenium_pek", "Woodwardia_uni")
gbFolder <- "/mnt/c/Users/Tanner/Documents/Research/ReFernment/ReFernment/examples/GB/"
gffFolder <- "/mnt/c/Users/Tanner/Documents/Research/ReFernment/ReFernment/examples/GFF/"
outputFolder <- "/mnt/c/Users/Tanner/Documents/Research/ReFernment/ReFernment/examples/"
ReFernment(gbFolder, gffFolder, outputFolder, genomes)

}
\author{
Tanner Robison
}
